+++ src/mainWindow.cpp
@@ -23,6 +23,11 @@
 #include <QFile>
 #include <QTextBrowser>
 
+#ifdef DEBUG
+#	include <QFileInfo>
+#	include <QDir>
+#endif
+
 #include "avdecc/helper.hpp"
 #include "avdecc/hiveLogItems.hpp"
 #include "internals/config.hpp"
@@ -148,6 +153,8 @@ void MainWindow::createMainToolBar()
 	interfaceLabel->setMinimumWidth(50);
 	_interfaceComboBox.setMinimumWidth(100);
 
+	_refreshControllerButton.setToolTip("Reload Controller");
+
 	auto* controllerEntityIDLabel = new QLabel("Controller ID: ");
 	controllerEntityIDLabel->setMinimumWidth(50);
 	_controllerEntityIDLabel.setMinimumWidth(100);
@@ -164,8 +171,17 @@ void MainWindow::createMainToolBar()
 
 	mainToolBar->addSeparator();
 
+	mainToolBar->addWidget(&_refreshControllerButton);
+
+	mainToolBar->addSeparator();
+
 	mainToolBar->addWidget(controllerEntityIDLabel);
 	mainToolBar->addWidget(&_controllerEntityIDLabel);
+
+#ifdef Q_OS_MAC
+	// See https://bugreports.qt.io/browse/QTBUG-13635
+	mainToolBar->setStyleSheet("QToolBar QLabel { padding-bottom: 5; }");
+#endif
 }
 
 void MainWindow::createControllerView()
@@ -258,6 +274,7 @@ void MainWindow::connectSignals()
 {
 	connect(&_protocolComboBox, QOverload<int>::of(&QComboBox::activated), this, &MainWindow::currentControllerChanged);
 	connect(&_interfaceComboBox, QOverload<int>::of(&QComboBox::activated), this, &MainWindow::currentControllerChanged);
+	connect(&_refreshControllerButton, &QPushButton::clicked, this, &MainWindow::currentControllerChanged);
 
 	connect(controllerTableView->selectionModel(), &QItemSelectionModel::currentChanged, this, &MainWindow::currentControlledEntityChanged);
 	connect(&_controllerDynamicHeaderView, &qt::toolkit::DynamicHeaderView::sectionChanged, this,
@@ -504,6 +521,21 @@ void MainWindow::connectSignals()
 
 	auto* refreshController = new QShortcut{ QKeySequence{ "Ctrl+R" }, this };
 	connect(refreshController, &QShortcut::activated, this, &MainWindow::currentControllerChanged);
+
+#ifdef DEBUG
+	auto* reloadStyleSheet = new QShortcut{ QKeySequence{ "F5" }, this };
+	connect(reloadStyleSheet, &QShortcut::activated, this,
+		[]()
+		{
+			// Load and apply the stylesheet
+			QFile styleFile{ QString{ RESOURCES_ROOT_DIR } + "/style.qss" };
+			if (styleFile.open(QFile::ReadOnly))
+			{
+				qApp->setStyleSheet(styleFile.readAll());
+				LOG_HIVE_DEBUG("StyleSheet reloaded");
+			}
+		});
+#endif
 }
 
 void MainWindow::showChangeLog(QString const title, QString const versionString)